impls/intersection.rs
