+++ src/impls.rs
@@ -1,2 +1,3 @@
 mod add;
 mod index;
+mod intersection;
